оптимизация отрисовки списка, удаление обработчиков событий у старых элементов
