+++ src/js/gallery.js
@@ -1,5 +1,7 @@
 'use strict';
 
+var Pic = require('./picture.js');
+
 var galleryContainer = document.querySelector('.gallery-overlay');
 var closeElement = galleryContainer.querySelector('.gallery-overlay-close');
 var imgElement = galleryContainer.querySelector('.gallery-overlay-image');
@@ -15,6 +17,7 @@ function Gallery() {
   this.hide = function(event) {
     event.preventDefault();
     galleryContainer.classList.add('invisible');
+    Pic.remove();
   };
 
   closeElement.addEventListener('click', this.hide);
