+++ src/js/picture.js
@@ -1,43 +1,23 @@
 'use strict';
 
+var getPictureElement = require('./getPictureElement.js');
 var gallery = require('./gallery.js');
 
-var template = document.querySelector('template');
-var templateContainer = 'content' in template ? template.content : template;
-var IMAGE_LOAD_TIMEOUT = 10000;
+var Picture = function(data, element) {
+  this.data = data;
+  this.element = getPictureElement(data, element);
 
-var getPictureElement = function(picture, index) {
-  var fotoBlock = templateContainer.querySelector('.picture').cloneNode(true);
-  fotoBlock.querySelector('.picture-comments').textContent = picture.comments;
-  fotoBlock.querySelector('.picture-likes').textContent = picture.likes;
+  var self = this;
 
-  var pic = fotoBlock.querySelector('img');
-
-  var contentImage = new Image(182, 182);
-  var imageTimeout = null;
-
-  contentImage.onload = function() {
-    clearTimeout(imageTimeout);
-    pic.src = picture.url;
-  };
-
-  contentImage.onerror = function() {
-    clearTimeout(imageTimeout);
-    fotoBlock.classList.add('picture-load-failure');
-  };
-
-  fotoBlock.addEventListener('click', function(event) {
+  this.element.addEventListener('click', function(event) {
     event.preventDefault();
-    gallery.show(index);
+    gallery.show(element);
   });
 
-  contentImage.src = picture.url;
-
-  imageTimeout = setTimeout(function() {
-    contentImage.classList.add('picture-load-failure');
-  }, IMAGE_LOAD_TIMEOUT);
-
-  return fotoBlock;
+  this.remove = function() {
+    self.element.removeEventListener('click', this.hide);
+  };
 };
 
-module.exports = getPictureElement;
+
+module.exports = Picture;
